+++ CHANGELOG.md
@@ -3,6 +3,44 @@ All notable changes to this project will be documented in this file.
 Format of this file follows [these](http://keepachangelog.com/) guidelines.
 This project adheres to [Semantic Versioning](http://semver.org/).
 
+## [0.7.0] - [2016-xx-xx]
+
+### Added
+
+- added proxy support to ```spiderhost.py``` to support use of anonymity networks like:
+    - [Luminati](https://luminati.io/)
+    - [Crawlera](http://crawlera.com/)
+    - [WonderProxy](https://wonderproxy.com/)
+    - [Distributed Scraping With Multiple Tor Circuits](http://blog.databigbang.com/tag/crawling-2/)
+
+```
+>Usage: spiderhost.py <spider> [<arg1> ... <argN>]
+
+spider hosts accept the name of a spider, the arguments to run the spider and
+optionally proxy server details. armed with all this info the spider host runs
+a spider and dumps the result to stdout.
+
+Options:
+  -h, --help            show this help message and exit
+  --log=LOGGING_LEVEL   logging level
+                        [DEBUG,INFO,WARNING,ERROR,CRITICAL,FATAL] - default =
+                        ERROR
+  --proxy=PROXY         proxy - default = None
+  --proxy-user=PROXY_USER
+                        proxy-user - default = None
+>
+>spiderhost.py --proxy=abc
+Usage: spiderhost.py <spider> [<arg1> ... <argN>]
+
+spiderhost.py: error: option --proxy: required format is host:port
+>
+>spiderhost.py --proxy-user=abc
+Usage: spiderhost.py <spider> [<arg1> ... <argN>]
+
+spiderhost.py: error: option --proxy-user: required format is user:password
+>
+```
+
 ## [0.6.0] - [2016-01-24]
 
 ### Changed